+++ lib/castle/commands/authenticate.rb
@@ -8,23 +8,15 @@ module Castle
       end
 
       def build(options = {})
-        validate!(options)
-        context = ContextMerger.call(@context, options[:context])
-        context = ContextSanitizer.call(context)
+        Castle::Validators::Present.call(options, %i[event user_id])
+        context = Castle::Context::Merger.call(@context, options[:context])
+        context = Castle::Context::Sanitizer.call(context)
 
-        Castle::Command.new('authenticate',
-                            options.merge(context: context,
-                                          sent_at: Castle::Utils::Timestamp.call),
-                            :post)
-      end
-
-      private
-
-      def validate!(options)
-        %i[event user_id].each do |key|
-          next unless options[key].to_s.empty?
-          raise Castle::InvalidParametersError, "#{key} is missing or empty"
-        end
+        Castle::Command.new(
+          'authenticate',
+          options.merge(context: context, sent_at: Castle::Utils::Timestamp.call),
+          :post
+        )
       end
     end
   end