+++ lib/castle/commands/identify.rb
@@ -8,28 +8,16 @@ module Castle
       end
 
       def build(options = {})
-        validate!(options)
-        context = ContextMerger.call(@context, options[:context])
-        context = ContextSanitizer.call(context)
+        Castle::Validators::Present.call(options, %i[user_id])
+        Castle::Validators::NotSupported.call(options, %i[properties])
+        context = Castle::Context::Merger.call(@context, options[:context])
+        context = Castle::Context::Sanitizer.call(context)
 
-        Castle::Command.new('identify',
-                            options.merge(context: context,
-                                          sent_at: Castle::Utils::Timestamp.call),
-                            :post)
-      end
-
-      private
-
-      def validate!(options)
-        %i[user_id].each do |key|
-          next unless options[key].to_s.empty?
-          raise Castle::InvalidParametersError, "#{key} is missing or empty"
-        end
-
-        if options[:properties]
-          raise Castle::InvalidParametersError,
-                'properties are not supported in identify calls'
-        end
+        Castle::Command.new(
+          'identify',
+          options.merge(context: context, sent_at: Castle::Utils::Timestamp.call),
+          :post
+        )
       end
     end
   end