+++ lib/castle/commands/impersonate.rb
@@ -9,31 +9,17 @@ module Castle
       end
 
       def build(options = {})
-        validate!(options)
-        context = ContextMerger.call(@context, options[:context])
-        context = ContextSanitizer.call(context)
+        Castle::Validators::Present.call(options, %i[user_id])
+        context = Castle::Context::Merger.call(@context, options[:context])
+        context = Castle::Context::Sanitizer.call(context)
 
-        validate_context!(context)
+        Castle::Validators::Present.call(context, %i[user_agent ip])
 
-        Castle::Command.new('impersonate',
-                            options.merge(context: context),
-                            :post)
-      end
-
-      private
-
-      def validate!(options)
-        %i[user_id].each do |key|
-          next unless options[key].to_s.empty?
-          raise Castle::InvalidParametersError, "#{key} is missing or empty"
-        end
-      end
-
-      def validate_context!(context)
-        %i[user_agent ip].each do |key|
-          next unless context[key].to_s.empty?
-          raise Castle::InvalidParametersError, "#{key} is missing or empty"
-        end
+        Castle::Command.new(
+          'impersonate',
+          options.merge(context: context),
+          :post
+        )
       end
     end
   end