+++ lib/castle/commands/review.rb
@@ -3,10 +3,11 @@
 module Castle
   module Commands
     class Review
-      def build(review_id)
-        raise Castle::InvalidParametersError if review_id.nil? || review_id.to_s.empty?
-
-        Castle::Command.new("reviews/#{review_id}", nil, :get)
+      class << self
+        def build(review_id)
+          Castle::Validators::Present.call({ review_id: review_id }, %i[review_id])
+          Castle::Command.new("reviews/#{review_id}", nil, :get)
+        end
       end
     end
   end