+++ spec/lib/castle/commands/authenticate_spec.rb
@@ -15,10 +15,10 @@ describe Castle::Commands::Authenticate do
   describe '.build' do
     subject(:command) { instance.build(payload) }
 
-    context 'simple merger' do
-      let(:payload) { default_payload.merge({ context: { test: { test2: '1' } } }) }
+    context 'with simple merger' do
+      let(:payload) { default_payload.merge(context: { test: { test2: '1' } }) }
       let(:command_data) do
-        default_payload.merge({ context: { test: { test1: '1', test2: '1' } } })
+        default_payload.merge(context: { test: { test1: '1', test2: '1' } })
       end
 
       it { expect(command.method).to be_eql(:post) }
@@ -26,10 +26,10 @@ describe Castle::Commands::Authenticate do
       it { expect(command.data).to be_eql(command_data) }
     end
 
-    context 'properties' do
-      let(:payload) { default_payload.merge({ properties: { test: '1' } }) }
+    context 'with properties' do
+      let(:payload) { default_payload.merge(properties: { test: '1' }) }
       let(:command_data) do
-        default_payload.merge({ properties: { test: '1' }, context: context })
+        default_payload.merge(properties: { test: '1' }, context: context)
       end
 
       it { expect(command.method).to be_eql(:post) }
@@ -37,10 +37,10 @@ describe Castle::Commands::Authenticate do
       it { expect(command.data).to be_eql(command_data) }
     end
 
-    context 'traits' do
-      let(:payload) { default_payload.merge({ traits: { test: '1' } }) }
+    context 'with traits' do
+      let(:payload) { default_payload.merge(traits: { test: '1' }) }
       let(:command_data) do
-        default_payload.merge({ traits: { test: '1' }, context: context })
+        default_payload.merge(traits: { test: '1' }, context: context)
       end
 
       it { expect(command.method).to be_eql(:post) }
@@ -48,10 +48,10 @@ describe Castle::Commands::Authenticate do
       it { expect(command.data).to be_eql(command_data) }
     end
 
-    context 'active true' do
-      let(:payload) { default_payload.merge({ context: { active: true } }) }
+    context 'when active true' do
+      let(:payload) { default_payload.merge(context: { active: true }) }
       let(:command_data) do
-        default_payload.merge({ context: context.merge(active: true) })
+        default_payload.merge(context: context.merge(active: true))
       end
 
       it { expect(command.method).to be_eql(:post) }
@@ -59,10 +59,10 @@ describe Castle::Commands::Authenticate do
       it { expect(command.data).to be_eql(command_data) }
     end
 
-    context 'active false' do
-      let(:payload) { default_payload.merge({ context: { active: false } }) }
+    context 'when active false' do
+      let(:payload) { default_payload.merge(context: { active: false }) }
       let(:command_data) do
-        default_payload.merge({ context: context.merge(active: false) })
+        default_payload.merge(context: context.merge(active: false))
       end
 
       it { expect(command.method).to be_eql(:post) }
@@ -70,9 +70,9 @@ describe Castle::Commands::Authenticate do
       it { expect(command.data).to be_eql(command_data) }
     end
 
-    context 'active string' do
-      let(:payload) { default_payload.merge({ context: { active: 'string' } }) }
-      let(:command_data) { default_payload.merge({ context: context }) }
+    context 'when active string' do
+      let(:payload) { default_payload.merge(context: { active: 'string' }) }
+      let(:command_data) { default_payload.merge(context: context) }
 
       it { expect(command.method).to be_eql(:post) }
       it { expect(command.path).to be_eql('authenticate') }
@@ -83,7 +83,7 @@ describe Castle::Commands::Authenticate do
   describe '#validate!' do
     subject(:validate!) { instance.build(payload) }
 
-    context 'event not present' do
+    context 'with event not present' do
       let(:payload) { {} }
 
       it do
@@ -93,7 +93,7 @@ describe Castle::Commands::Authenticate do
       end
     end
 
-    context 'user_id not present' do
+    context 'with user_id not present' do
       let(:payload) { { event: '$login.track' } }
 
       it do
@@ -103,7 +103,7 @@ describe Castle::Commands::Authenticate do
       end
     end
 
-    context 'event and user_id present' do
+    context 'with event and user_id present' do
       let(:payload) { { event: '$login.track', user_id: '1234' } }
 
       it { expect { validate! }.not_to raise_error }