+++ spec/lib/castle/context/merger_spec.rb
@@ -1,20 +1,20 @@
 # frozen_string_literal: true
 
-describe Castle::ContextMerger do
+describe Castle::Context::Merger do
   let(:first) { { test: { test1: { c: '4' } } } }
 
-  context '#call' do
+  describe '#call' do
     subject { described_class.call(first, second) }
 
     let(:result) { { test: { test1: { c: '4', d: '5' } } } }
 
-    context 'symbol keys' do
+    context 'with symbol keys' do
       let(:second) { { test: { test1: { d: '5' } } } }
 
       it { is_expected.to eq(result) }
     end
 
-    context 'string keys' do
+    context 'with string keys' do
       let(:second) { { 'test' => { 'test1' => { 'd' => '5' } } } }
 
       it { is_expected.to eq(result) }